--- .github/workflows/deploy-prod-v2.yml
@@ -1,38 +1,0 @@
-name: Deploy function to dev
-
-on:
-  push:
-    tags:
-      - v*
-
-jobs:
-  build-and-deploy:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: "read"
-      id-token: "write"
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          ref: main # The branch to checkout for the deploy
-          persist-credentials: false
-
-      - id: "auth"
-        uses: "google-github-actions/auth@v1"
-        with:
-          credentials_json: "${{ secrets.GCLOUD_SERVICE_ACCOUNT }}"
-
-      - name: "Set up Cloud SDK"
-        uses: "google-github-actions/setup-gcloud@v1"
-
-      - name: Deploy function
-        run: |
-          gcloud functions deploy create_tree_v2 \
-            --project coreofscience \
-            --trigger-event "providers/cloud.firestore/eventTypes/document.create" \
-            --trigger-resource "projects/coreofscience/databases/(default)/documents/trees/{treeId}" \
-            --set-env-vars STORAGEBUCKET="coreofscience.appspot.com",DATABASEURL="https://coreofscience.firebaseio.com" \
-            --source ./create-tree-v2 \
-            --entry-point create_tree_v2 \
-            --runtime python310
