--- create-tree-v2/requirements.txt
@@ -1,4 +1,0 @@
-python-sap==2.0.0
-firebase-admin==4.4.0
-google-cloud-storage==1.31.2
-google-cloud-logging==1.15.1
